+++ crates/serialize/src/lib.rs
@@ -1,0 +1,8 @@
+mod i64;
+mod serialize;
+mod string;
+mod u64;
+mod vec;
+mod vec_u8;
+
+pub use serialize::Serialize;
