+++ src/graph/dictionary_data_memory.rs
@@ -1,9 +1,9 @@
 use super::dictionary_data::DictionaryData;
 use super::dictionary_value::DictionaryValue;
 use crate::storage::HashMultiMap;
-use crate::storage::Serialize;
 use crate::storage::StableHash;
 use crate::DbError;
+use serialize::Serialize;
 
 pub(crate) struct DictionaryDataMemory<T>
 where
