+++ src/graph/dictionary_data_storage_indexes.rs
@@ -1,4 +1,4 @@
-use crate::storage::Serialize;
+use serialize::Serialize;
 
 pub(super) struct DictionaryDataStorageIndexes {
     pub(super) index: i64,
