+++ src/storage/hash_map_data_memory.rs
@@ -1,9 +1,9 @@
 use super::hash_map_data::HashMapData;
 use super::hash_map_key_value::HashMapKeyValue;
 use super::hash_map_meta_value::HashMapMetaValue;
-use super::Serialize;
 use super::StableHash;
 use crate::DbError;
+use serialize::Serialize;
 use std::hash::Hash;
 
 pub(crate) struct HashMapDataMemory<K, T>
