+++ src/storage/hash_multi_map.rs
@@ -2,8 +2,8 @@ use super::hash_map_data_memory::HashMapDataMemory;
 use super::hash_map_impl::HashMapImpl;
 use super::hash_map_key_value::HashMapKeyValue;
 use super::hash_multi_map_impl::HashMultiMapImpl;
-use super::Serialize;
 use super::StableHash;
+use serialize::Serialize;
 use std::hash::Hash;
 
 pub(crate) type HashMultiMap<K, T> = HashMultiMapImpl<K, T, HashMapDataMemory<K, T>>;
