+++ src/storage/storage_hash_multi_map.rs
@@ -4,12 +4,12 @@ use super::hash_map_data_storage::HashMapDataStorage;
 use super::hash_map_impl::HashMapImpl;
 use super::hash_multi_map::HashMultiMap;
 use super::hash_multi_map_impl::HashMultiMapImpl;
-use super::serialize::Serialize;
 use super::stable_hash::StableHash;
 use super::storage_data::StorageData;
 use super::Storage;
 use super::StorageHashMap;
 use crate::DbError;
+use serialize::Serialize;
 use std::hash::Hash;
 
 pub(crate) type StorageHashMultiMap<K, T, Data = FileStorageData> =
